pyproject.toml
[project]
name = "donotes"
version = "0.1.0"
description = "AI-powered personal productivity bot: capture voice/text via Telegram, auto-transcribe, extract actions, create calendar events, and send email digests"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=21.0",
    "openai>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.0",
    "jinja2>=3.1.0",
    "thefuzz>=0.22.0",
    "python-Levenshtein>=0.25.0",
    "eval-type-backport>=0.1.3; python_version < '3.10'",
]

[project.optional-dependencies]
menubar = ["rumps>=0.4.0", "py2app>=0.28.0"]
mlx = ["lightning-whisper-mlx>=0.0.10"]
dev = ["pytest", "pytest-asyncio", "ruff", "mypy"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "src*"]
